split lakectl cmds - part 3 (#6277)
* lakectl cmd part 2

* lakectl cmd split part 3

* update cli docs
