+++ cmd/lakectl/cmd/branch_delete.go
@@ -1,0 +1,34 @@
+package cmd
+
+import (
+	"fmt"
+	"net/http"
+
+	"github.com/spf13/cobra"
+)
+
+var branchDeleteCmd = &cobra.Command{
+	Use:               "delete <branch uri>",
+	Short:             "Delete a branch in a repository, along with its uncommitted changes (CAREFUL)",
+	Example:           "lakectl branch delete lakefs://example-repo/example-branch",
+	Args:              cobra.ExactArgs(1),
+	ValidArgsFunction: ValidArgsRepository,
+	Run: func(cmd *cobra.Command, args []string) {
+		confirmation, err := Confirm(cmd.Flags(), "Are you sure you want to delete branch")
+		if err != nil || !confirmation {
+			Die("Delete branch aborted", 1)
+		}
+		client := getClient()
+		u := MustParseBranchURI("branch", args[0])
+		fmt.Printf("Branch: %s\n", u)
+		resp, err := client.DeleteBranchWithResponse(cmd.Context(), u.Repository, u.Ref)
+		DieOnErrorOrUnexpectedStatusCode(resp, err, http.StatusNoContent)
+	},
+}
+
+//nolint:gochecknoinits
+func init() {
+	AssignAutoConfirmFlag(branchDeleteCmd.Flags())
+
+	branchCmd.AddCommand(branchDeleteCmd)
+}
