+++ cmd/lakectl/cmd/branch_reset.go
@@ -1,0 +1,75 @@
+package cmd
+
+import (
+	"fmt"
+	"net/http"
+
+	"github.com/spf13/cobra"
+	"github.com/treeverse/lakefs/pkg/api"
+)
+
+// lakectl branch reset lakefs://myrepo/main --commit commitId --prefix path --object path
+var branchResetCmd = &cobra.Command{
+	Use:     "reset <branch uri> [--prefix|--object]",
+	Example: "lakectl branch reset lakefs://example-repo/example-branch",
+	Short:   "Reset uncommitted changes - all of them, or by path",
+	Long: `reset changes.  There are four different ways to reset changes:
+  1. reset all uncommitted changes - reset lakefs://myrepo/main 
+  2. reset uncommitted changes under specific path - reset lakefs://myrepo/main --prefix path
+  3. reset uncommitted changes for specific object - reset lakefs://myrepo/main --object path`,
+	Args:              cobra.ExactArgs(1),
+	ValidArgsFunction: ValidArgsRepository,
+	Run: func(cmd *cobra.Command, args []string) {
+		clt := getClient()
+		u := MustParseBranchURI("branch", args[0])
+		fmt.Printf("Branch: %s\n", u)
+		prefix, err := cmd.Flags().GetString("prefix")
+		if err != nil {
+			DieErr(err)
+		}
+		object, err := cmd.Flags().GetString("object")
+		if err != nil {
+			DieErr(err)
+		}
+
+		var reset api.ResetCreation
+		var confirmationMsg string
+		switch {
+		case len(prefix) > 0:
+			confirmationMsg = fmt.Sprintf("Are you sure you want to reset all uncommitted changes from path: %s", prefix)
+			reset = api.ResetCreation{
+				Path: &prefix,
+				Type: "common_prefix",
+			}
+		case len(object) > 0:
+			confirmationMsg = fmt.Sprintf("Are you sure you want to reset all uncommitted changes for object: %s", object)
+			reset = api.ResetCreation{
+				Path: &object,
+				Type: "object",
+			}
+		default:
+			confirmationMsg = "Are you sure you want to reset all uncommitted changes"
+			reset = api.ResetCreation{
+				Type: "reset",
+			}
+		}
+
+		confirmation, err := Confirm(cmd.Flags(), confirmationMsg)
+		if err != nil || !confirmation {
+			Die("Reset aborted", 1)
+			return
+		}
+		resp, err := clt.ResetBranchWithResponse(cmd.Context(), u.Repository, u.Ref, api.ResetBranchJSONRequestBody(reset))
+		DieOnErrorOrUnexpectedStatusCode(resp, err, http.StatusNoContent)
+	},
+}
+
+//nolint:gochecknoinits
+func init() {
+	AssignAutoConfirmFlag(branchResetCmd.Flags())
+
+	branchResetCmd.Flags().String("prefix", "", "prefix of the objects to be reset")
+	branchResetCmd.Flags().String("object", "", "path to object to be reset")
+
+	branchCmd.AddCommand(branchResetCmd)
+}
